Part 2 - Update 2
Added Bi-directional Relationship between Car and Owner to demonstrate how it could be handled when serialized in a Json file with Jackson framework.
Created Two Deserialize method because i could not wrap my head around how to create one in a elegant way.
Working with @JsonManagedReference & @JsonBackReference to handle the Many-to-One relationship so we dont get Infinite recursion. + modified the toString also.
Owner owns this relationship of Cars.

Annnnddd. Nope it's not Tested..

- Created Owner out of AppUser
- Added Relationship Between Owner and Car
- Made sure no infinite recursion happened.